datalake-ingester/requirements.txt
boto3==1.34.0
botocore==1.34.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pymongo==4.6.1
cryptography==41.0.7
python-dotenv==1.0.0